+++ src/acme_portal_sdk/deployment_finder.py
@@ -1,9 +1,8 @@
 from abc import ABC, abstractmethod
 from dataclasses import asdict, dataclass, field
-from typing import Any, Dict, List, Optional, TYPE_CHECKING
+from typing import Any, Dict, List, Optional
 
-if TYPE_CHECKING:
-    from .flow_finder import FlowDetails
+from .flow_finder import FlowDetails
 
 
 @dataclass
@@ -58,19 +57,37 @@ class DeploymentFinder(ABC):
     @abstractmethod
     def get_deployments(
         self,
+        *,
         deployments_to_fetch: Optional[List[DeploymentDetails]] = None,
-        flows_to_fetch: Optional[List["FlowDetails"]] = None
+        flows_to_fetch: Optional[List["FlowDetails"]] = None,
     ) -> List[DeploymentDetails]:
         """Method to find deployments, to be implemented by subclasses.
-        
-        Args:
+
+        kwargs:
             deployments_to_fetch: Optional list of specific deployments to re-fetch
             flows_to_fetch: Optional list of flows to re-fetch deployments for
-            
+
         Returns:
             List of DeploymentDetails objects
         """
         pass
 
-    def __call__(self) -> List[Dict[str, Any]]:
-        return [x.to_dict() for x in self.get_deployments()]
+    def __call__(
+        self,
+        *,
+        deployments_to_fetch: Optional[List[dict]] = None,
+        flows_to_fetch: Optional[List[dict]] = None,
+    ) -> List[Dict[str, Any]]:
+        return [
+            x.to_dict()
+            for x in self.get_deployments(
+                deployments_to_fetch=[
+                    DeploymentDetails.from_dict(x) for x in deployments_to_fetch
+                ]
+                if deployments_to_fetch is not None
+                else deployments_to_fetch,
+                flows_to_fetch=[FlowDetails.from_dict(x) for x in flows_to_fetch]
+                if flows_to_fetch is not None
+                else flows_to_fetch,
+            )
+        ]