+++ src/acme_portal_sdk/flow_finder.py
@@ -18,8 +18,8 @@ class FlowDetails:
         source_relative: Relative path to the source file from some known root
         line_number: Line number where the flow is defined in the source file
         grouping: Desired grouping of the flow in the context of the project (for navigation)
-        child_attributes: Additional attributes specific to implementation (e.g., obj_type, obj_name, 
-                         obj_parent_type, obj_parent, module, import_path for Prefect). Should not be 
+        child_attributes: Additional attributes specific to implementation (e.g., obj_type, obj_name,
+                         obj_parent_type, obj_parent, module, import_path for Prefect). Should not be
                          set by subclasses, but may be set by users to add custom information.
     """
 
@@ -48,20 +48,35 @@ class FlowFinder(ABC):
 
     @abstractmethod
     def find_flows(
-        self, 
+        self,
+        *,
         flows_to_fetch: Optional[List[FlowDetails]] = None,
-        flow_groups: Optional[List[str]] = None
+        flow_groups: Optional[List[str]] = None,
     ) -> List[FlowDetails]:
         """Method to find flows, to be implemented by subclasses.
-        
-        Args:
+
+        kwargs:
             flows_to_fetch: Optional list of flows to selectively re-fetch data for
             flow_groups: Optional list of flow group names to selectively re-fetch
-            
+
         Returns:
             List of FlowDetails objects
         """
         pass
 
-    def __call__(self) -> List[Dict[str, Any]]:
-        return [x.to_dict() for x in self.find_flows()]
+    def __call__(
+        self,
+        *,
+        flows_to_fetch: Optional[List[dict]] = None,
+        flow_groups: Optional[List[str]] = None,
+    ) -> List[Dict[str, Any]]:
+        """Passthrough method for find_flows call that is operating on JSON (serializable) types"""
+        return [
+            x.to_dict()
+            for x in self.find_flows(
+                flows_to_fetch=[FlowDetails.from_dict(x) for x in flows_to_fetch]
+                if flows_to_fetch is not None
+                else flows_to_fetch,
+                flow_groups=flow_groups,
+            )
+        ]